+++ components/shared_code/src/shared_data_model/sources/junit.py
@@ -4,7 +4,7 @@ from pydantic import HttpUrl
 
 from shared_data_model.meta.entity import Color, Entity, EntityAttribute, EntityAttributeType
 from shared_data_model.meta.source import Source
-from shared_data_model.parameters import TestResult, access_parameters
+from shared_data_model.parameters import TestResult, TestResultAggregationStrategy, access_parameters
 
 ALL_JUNIT_METRICS = ["source_up_to_dateness", "test_cases", "test_suites", "tests"]
 
@@ -31,6 +31,7 @@ JUNIT = Source(
     url=HttpUrl("https://junit.org/junit5/"),
     parameters={
         "test_result": TestResult(metrics=["test_suites", "tests"], values=["errored", "failed", "passed", "skipped"]),
+        "test_result_aggregation_strategy": TestResultAggregationStrategy(),
         **access_parameters(ALL_JUNIT_METRICS, source_type="JUnit report", source_type_format="XML"),
     },
     entities={